requirements.txt
requests
python-decouple
apscheduler
